STR-483 Bridge client factory (#387)
* fntest/fw: moved factories to new module

* bridge-client: reworked some init strings

* bridge-client: added bridge clients to envs but it's broken

* tests: fix btcio_inscriber.py

* format: factory.py

* tests: fix sync_bicoin_reorg.py

* lint: python lints

* fix: waiting a little bit shorter in sync_bitcoin_reorg

---------
